Initial batch of changes for updated NSDI experiments
