+++ skyplane/cli/cli_impl/cp_replicate.py
@@ -33,13 +33,23 @@ def generate_topology(
     solve: bool,
     num_connections: int = 32,
     max_instances: int = 1,
+    solver_class: str = "ILP",
     solver_total_gbyte_to_transfer: Optional[float] = None,
     solver_required_throughput_gbits: float = 4,
     solver_throughput_grid: Optional[pathlib.Path] = skyplane_root / "profiles" / "throughput.csv",
     solver_verbose: Optional[bool] = False,
     args: Optional[Dict] = None,
 ) -> ReplicationTopology:
-    if solve:
+    if src_region == dst_region:  # intra-region transfer w/o solver
+        topo = ReplicationTopology()
+        for i in range(max_instances):
+            topo.add_objstore_instance_edge(src_region, src_region, i)
+            topo.add_instance_objstore_edge(src_region, i, src_region)
+        topo.cost_per_gb = 0
+        return topo
+    elif solve:
+        from skyplane.replicate.solver import ThroughputProblem
+
         if src_region == dst_region:
             e = "Solver is not supported for intra-region transfers, run without the --solve flag"
             typer.secho(e, fg="red", err=True)
@@ -52,13 +62,7 @@ def generate_topology(
                 client.report_usage_data("error", stats, destination)
 
             raise typer.Exit(1)
-
-        # build problem and solve
-        from skyplane.replicate.solver import ThroughputProblem
-        from skyplane.replicate.solver_ilp import ThroughputSolverILP
-
         assert solver_throughput_grid is not None and solver_total_gbyte_to_transfer is not None
-        tput = ThroughputSolverILP(solver_throughput_grid)
         problem = ThroughputProblem(
             src=src_region,
             dst=dst_region,
@@ -66,30 +70,38 @@ def generate_topology(
             gbyte_to_transfer=solver_total_gbyte_to_transfer,
             instance_limit=max_instances,
         )
-        with Timer() as t:
-            with console.status("Solving for the optimal transfer plan"):
-                solution = tput.solve_min_cost(
-                    problem, solver=ThroughputSolverILP.choose_solver(), solver_verbose=solver_verbose, save_lp_path=None
-                )
-        typer.secho(f"Solving for the optimal transfer plan took {t.elapsed:.2f}s", fg="green")
-        topo, scale_factor = tput.to_replication_topology(solution)
-        logger.fs.debug(f"Scaled solution by {scale_factor:.2f}x")
-        topo.cost_per_gb = solution.cost_egress / solution.problem.gbyte_to_transfer
-        return topo
-    else:
-        if src_region == dst_region:
-            topo = ReplicationTopology()
-            for i in range(max_instances):
-                topo.add_objstore_instance_edge(src_region, src_region, i)
-                topo.add_instance_objstore_edge(src_region, i, src_region)
-            topo.cost_per_gb = 0
+
+        if solver_class == "ILP":
+            from skyplane.replicate.solver_ilp import ThroughputSolverILP
+
+            tput = ThroughputSolverILP(solver_throughput_grid)
+            with Timer() as t:
+                with console.status("Solving for the optimal transfer plan"):
+                    solution = tput.solve_min_cost(
+                        problem, solver=ThroughputSolverILP.choose_solver(), solver_verbose=solver_verbose, save_lp_path=None
+                    )
+            typer.secho(f"Solving for the optimal transfer plan took {t.elapsed:.2f}s", fg="green")
+            topo, scale_factor = tput.to_replication_topology(solution)
+            logger.fs.debug(f"Scaled solution by {scale_factor:.2f}x")
+            topo.cost_per_gb = solution.cost_egress / solution.problem.gbyte_to_transfer
+            return topo
+        elif solver_class == "RON":
+            from skyplane.replicate.solver_ron import ThroughputSolverRON
+
+            tput = ThroughputSolverRON(solver_throughput_grid)
+            solution = tput.solve(problem)
+            topo, scale_factor = tput.to_replication_topology(solution)
+            topo.cost_per_gb = solution.cost_egress / solution.problem.gbyte_to_transfer
+            return topo
         else:
-            topo = ReplicationTopology()
-            for i in range(max_instances):
-                topo.add_objstore_instance_edge(src_region, src_region, i)
-                topo.add_instance_instance_edge(src_region, i, dst_region, i, num_connections)
-                topo.add_instance_objstore_edge(dst_region, i, dst_region)
-            topo.cost_per_gb = CloudProvider.get_transfer_cost(src_region, dst_region)
+            raise NotImplementedError(f"Solver class {solver_class} not implemented")
+    else:  # inter-region transfer w/o solver
+        topo = ReplicationTopology()
+        for i in range(max_instances):
+            topo.add_objstore_instance_edge(src_region, src_region, i)
+            topo.add_instance_instance_edge(src_region, i, dst_region, i, num_connections)
+            topo.add_instance_objstore_edge(dst_region, i, dst_region)
+        topo.cost_per_gb = CloudProvider.get_transfer_cost(src_region, dst_region)
         return topo
 
 
@@ -263,9 +275,7 @@ def launch_replication_job(
     # logging options
     time_limit_seconds: Optional[int] = None,
     log_interval_s: float = 1.0,
-    src_region: Optional[str] = None,
-    dst_region: Optional[str] = None,
-    args: Optional[Dict] = None,
+    error_reporting_args: Optional[Dict] = None,
 ):
     if "SKYPLANE_DOCKER_IMAGE" in os.environ:
         rprint(f"[bright_black]Using overridden docker image: {gateway_docker_image}[/bright_black]")
@@ -319,6 +329,7 @@ def launch_replication_job(
             job,
             show_spinner=True,
             log_interval_s=log_interval_s,
+            log_to_file=True,
             time_limit_seconds=time_limit_seconds,
             multipart=multipart_enabled,
             write_profile=debug,
@@ -341,7 +352,7 @@ def launch_replication_job(
         client = UsageClient()
         if client.enabled():
             error_dict = {"loc": "launch_replication_job", "message": str(e)[:150]}
-            err_stats = client.make_error(src_region, dst_region, error_dict, args)
+            err_stats = client.make_error(job.source_region, job.dest_region, error_dict, error_reporting_args)
             destination = client.write_usage_data(err_stats)
             client.report_usage_data("error", err_stats, destination)
         os._exit(1)  # exit now
@@ -360,7 +371,7 @@ def launch_replication_job(
         client = UsageClient()
         if client.enabled():
             error_dict = {"loc": "replication_monitor", "message": err[:150]}
-            err_stats = client.make_error(src_region, dst_region, error_dict, args)
+            err_stats = client.make_error(job.source_region, job.dest_region, error_dict, error_reporting_args)
             destination = client.write_usage_data(err_stats)
             client.report_usage_data("error", err_stats, destination)
         raise typer.Exit(1)
@@ -375,7 +386,7 @@ def launch_replication_job(
         client = UsageClient()
         if client.enabled():
             error_dict = {"loc": "replication_monitor", "message": stats.monitor_status}
-            err_stats = client.make_error(src_region, dst_region, error_dict, args)
+            err_stats = client.make_error(job.source_region, job.dest_region, error_dict, error_reporting_args)
             destination = client.write_usage_data(err_stats)
             client.report_usage_data("error", err_stats, destination)
     return stats
