+++ skyplane/replicate/solver_ilp.py
@@ -1,12 +1,5 @@
-import shutil
-from collections import namedtuple
-from typing import Tuple, List, Dict, Optional
-
 import cvxpy as cp
-import graphviz as gv
-import numpy as np
 
-from skyplane.replicate.replication_plan import ReplicationTopology
 from skyplane.replicate.solver import ThroughputSolver, ThroughputProblem, GBIT_PER_GBYTE, ThroughputSolution
 from skyplane.utils import logger
 
@@ -28,7 +21,7 @@ class ThroughputSolverILP(ThroughputSolver):
                 logger.fs.warning("Gurobi and CoinOR not installed, using GLPK instead.")
                 return cp.GLPK
 
-    def solve_min_cost(self, p: ThroughputProblem, solver=cp.GLPK, solver_verbose=False, save_lp_path=None):
+    def solve_min_cost(self, p: ThroughputProblem, solver=cp.GLPK, solver_verbose=False, save_lp_path=None) -> ThroughputSolution:
         regions = self.get_regions()
         sources = [regions.index(p.src)]
         sinks = [regions.index(p.dst)]
@@ -148,166 +141,3 @@ class ThroughputSolverILP(ThroughputSolver):
                 baseline_cost_instance=baseline_instance_cost,
                 baseline_cost_total=baseline_egress_cost + baseline_instance_cost,
             )
-
-    def print_solution(self, solution: ThroughputSolution):
-        if solution.is_feasible:
-            regions = self.get_regions()
-            logger.debug(
-                f"Total cost: ${solution.cost_total:.4f} (egress: ${solution.cost_egress:.4f}, instance: ${solution.cost_instance:.4f})"
-            )
-            logger.debug(f"Total throughput: [{', '.join(str(round(t, 2)) for t in solution.throughput_achieved_gbits)}] Gbps")
-            logger.debug(f"Total runtime: {solution.transfer_runtime_s:.2f}s")
-            region_inst_count = {regions[i]: int(solution.var_instances_per_region[i]) for i in range(len(regions))}
-            logger.debug("Instance regions: [{}]".format(", ".join(f"{r}={c}" for r, c in region_inst_count.items() if c > 0)))
-            logger.debug("Flow matrix:")
-            for i, src in enumerate(regions):
-                for j, dst in enumerate(regions):
-                    if solution.var_edge_flow_gigabits[i, j] > 0:
-                        gb_sent = solution.transfer_runtime_s * solution.var_edge_flow_gigabits[i, j] * GBIT_PER_GBYTE
-                        s = f"\t{src} -> {dst}: {solution.var_edge_flow_gigabits[i, j]:.2f} Gbps with {solution.var_conn[i, j]:.1f} connections, "
-                        s += f"{gb_sent:.1f}GB (link capacity = {solution.problem.const_throughput_grid_gbits[i, j]:.2f} Gbps)"
-                        logger.debug(s)
-        else:
-            logger.debug("No feasible solution")
-
-    def plot_graphviz(self, solution: ThroughputSolution) -> Optional[gv.Digraph]:
-        # if dot is not installed
-        has_dot = shutil.which("dot") is not None
-        if not has_dot:
-            logger.error("Graphviz is not installed. Please install it to plot the solution (sudo apt install graphviz).")
-            return None
-
-        regions = self.get_regions()
-        g = gv.Digraph(name="throughput_graph")
-        g.attr(rankdir="LR")
-
-        label = f"{solution.problem.src} to {solution.problem.dst}\n"
-        label += f"[{', '.join(str(round(t, 2)) for t in solution.throughput_achieved_gbits)}] Gbps, ${solution.cost_total:.4f}\n"
-        label += f"Average: ${solution.cost_total / solution.transfer_runtime_s:.4f}/GB"
-        g.attr(label=label)
-        g.attr(labelloc="t")
-        for i, src in enumerate(regions):
-            for j, dst in enumerate(regions):
-                if solution.var_edge_flow_gigabits[i, j] > 0:
-                    link_cost = self.get_path_cost(src, dst)
-                    label = f"{solution.var_edge_flow_gigabits[i, j]:.2f} Gbps (of {solution.problem.const_throughput_grid_gbits[i, j]:.2f}Gbps), "
-                    label += f"\n${link_cost:.4f}/GB over {solution.var_conn[i, j]:.1f}c"
-                    src_label = f"{src.replace(':', '/')}x{solution.var_instances_per_region[i]:.1f}"
-                    dst_label = f"{dst.replace(':', '/')}x{solution.var_instances_per_region[j]:.1f}"
-                    g.edge(src_label, dst_label, label=label)
-        return g
-
-    def to_replication_topology(self, solution: ThroughputSolution, scale_to_capacity=True) -> Tuple[ReplicationTopology, float]:
-        regions = self.get_regions()
-        Edge = namedtuple("Edge", ["src_region", "src_instance_idx", "dst_region", "dst_instance_idx", "connections"])
-
-        # compute connections to target per instance
-        ninst = solution.var_instances_per_region
-        average_egress_conns = [0 if ninst[i] == 0 else np.ceil(solution.var_conn[i, :].sum() / ninst[i]) for i in range(len(regions))]
-        average_ingress_conns = [0 if ninst[i] == 0 else np.ceil(solution.var_conn[:, i].sum() / ninst[i]) for i in range(len(regions))]
-
-        # first assign source instances to destination regions
-        src_edges: List[Edge] = []
-        n_instances = {}
-        for i, src in enumerate(regions):
-            src_instance_idx, src_instance_connections = 0, 0
-            for j, dst in enumerate(regions):
-                if solution.var_edge_flow_gigabits[i, j] > 0:
-                    connections_to_allocate = np.rint(solution.var_conn[i, j]).astype(int)
-                    while connections_to_allocate > 0:
-                        # if this edge would exceed the instance connection limit, partially add connections to
-                        # current instance and increment instance
-                        if connections_to_allocate + src_instance_connections > average_egress_conns[i]:
-                            partial_conn = average_egress_conns[i] - src_instance_connections
-                            connections_to_allocate -= partial_conn
-                            assert connections_to_allocate >= 0, f"connections_to_allocate = {connections_to_allocate}"
-                            assert partial_conn >= 0, f"partial_conn = {partial_conn}"
-                            if partial_conn > 0:
-                                src_edges.append(Edge(src, src_instance_idx, dst, None, partial_conn))
-                                logger.fs.warning(
-                                    f"{src}:{src_instance_idx}:{src_instance_connections}c -> {dst} (partial): {partial_conn}c of {connections_to_allocate}c remaining"
-                                )
-                            src_instance_idx += 1
-                            src_instance_connections = 0
-                        else:
-                            partial_conn = connections_to_allocate
-                            connections_to_allocate = 0
-                            src_edges.append(Edge(src, src_instance_idx, dst, None, partial_conn))
-                            logger.fs.warning(
-                                f"{src}:{src_instance_idx}:{src_instance_connections}c -> {dst}: {partial_conn}c of {connections_to_allocate}c remaining"
-                            )
-                            src_instance_connections += partial_conn
-            n_instances[i] = src_instance_idx + 1
-
-        # assign destination instances (currently None) to Edges
-        dst_edges = []
-        dsts_instance_idx = {i: 0 for i in regions}
-        dsts_instance_conn = {i: 0 for i in regions}
-        for e in src_edges:
-            connections_to_allocate = np.rint(e.connections).astype(int)
-            while connections_to_allocate > 0:
-                if connections_to_allocate + dsts_instance_conn[e.dst_region] > average_ingress_conns[regions.index(e.dst_region)]:
-                    partial_conn = average_ingress_conns[regions.index(e.dst_region)] - dsts_instance_conn[e.dst_region]
-                    connections_to_allocate = connections_to_allocate - partial_conn
-                    if partial_conn > 0:
-                        dst_edges.append(
-                            Edge(e.src_region, e.src_instance_idx, e.dst_region, dsts_instance_idx[e.dst_region], partial_conn)
-                        )
-                        logger.fs.warning(
-                            f"{e.src_region}:{e.src_instance_idx}:{dsts_instance_conn[e.dst_region]}c -> {e.dst_region}:{dsts_instance_idx[e.dst_region]}:{dsts_instance_conn[e.dst_region]}c (partial): {partial_conn}c of {connections_to_allocate}c remaining"
-                        )
-                    dsts_instance_idx[e.dst_region] += 1
-                    dsts_instance_conn[e.dst_region] = 0
-                else:
-                    dst_edges.append(
-                        Edge(e.src_region, e.src_instance_idx, e.dst_region, dsts_instance_idx[e.dst_region], connections_to_allocate)
-                    )
-                    logger.fs.warning(
-                        f"{e.src_region}:{e.src_instance_idx}:{dsts_instance_conn[e.dst_region]}c -> {e.dst_region}:{dsts_instance_idx[e.dst_region]}:{dsts_instance_conn[e.dst_region]}c: {connections_to_allocate}c remaining"
-                    )
-                    dsts_instance_conn[e.dst_region] += connections_to_allocate
-                    connections_to_allocate = 0
-
-        # scale connections up to saturate links
-        if scale_to_capacity:
-            conns_egress: Dict[Tuple[str, int], int] = {}
-            conns_ingress: Dict[Tuple[str, int], int] = {}
-            for e in dst_edges:
-                conns_egress[(e.src_region, e.src_instance_idx)] = e.connections + conns_egress.get((e.src_region, e.src_instance_idx), 0)
-                conns_ingress[(e.dst_region, e.dst_instance_idx)] = e.connections + conns_ingress.get((e.dst_region, e.dst_instance_idx), 0)
-            bottleneck_capacity = max(list(conns_egress.values()) + list(conns_ingress.values()))
-            scale_factor = 64 / bottleneck_capacity
-            logger.fs.warning(f"Scaling connections by {scale_factor:.2f}x")
-            dst_edges = [e._replace(connections=int(e.connections * scale_factor)) for e in dst_edges]
-        else:
-            scale_factor = 1.0
-
-        # build ReplicationTopology
-        obj_store_edges = set()
-        replication_topology = ReplicationTopology()
-        for e in dst_edges:
-            if e.connections >= 1:
-                # connect source to destination
-                replication_topology.add_instance_instance_edge(
-                    src_region=e.src_region,
-                    src_instance=e.src_instance_idx,
-                    dest_region=e.dst_region,
-                    dest_instance=e.dst_instance_idx,
-                    num_connections=e.connections,
-                )
-
-                # connect source instances to source gateway
-                if e.src_region == solution.problem.src and ("src", e.src_region, e.src_instance_idx) not in obj_store_edges:
-                    replication_topology.add_objstore_instance_edge(
-                        src_region=e.src_region, dest_region=e.src_region, dest_instance=e.src_instance_idx
-                    )
-                    obj_store_edges.add(("src", e.src_region, e.src_instance_idx))
-
-                # connect destination instances to destination gateway
-                if e.dst_region == solution.problem.dst and ("dst", e.dst_region, e.dst_instance_idx) not in obj_store_edges:
-                    replication_topology.add_instance_objstore_edge(
-                        src_region=e.dst_region, src_instance=e.dst_instance_idx, dest_region=e.dst_region
-                    )
-                    obj_store_edges.add(("dst", e.dst_region, e.dst_instance_idx))
-
-        return replication_topology, scale_factor
